fix: add validation to advisers in UpdateTransactionsController and change img src in index view in Cards folder

+++ app/Http/Controllers/Users/UpdateUsersController.php
@@ -76,7 +76,9 @@ class UpdateUsersController extends Controller
                     $shopkeeperAdviser->save();
                 }
             } else {
-                ShopkeeperAdviser::destroy($shopkeeperAdviser->id);
+                if (isset($shopkeeperAdviser->id)) {
+                    ShopkeeperAdviser::destroy($shopkeeperAdviser->id);
+                }
             }
         }
